requirements.txt
requests
python-valve
python-a2s
reactivex
influxdb-client